Allow for custom pi-greeter.conf
In order to make the pi-greeter.conf customizable through external debian
packages, the config file is created as symlink via update-alternatives. The
default config is now installed to /usr/share/pi-greeter and linked by default
with a priority of 50.

This will allow OEMs and other users, who want to customize the greeter to ship
custom config files.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>